requirements.txt
Django==1.5.8
South==0.8.4
Pillow==2.3.1
Fabric==1.8.3
fabric-virtualenv==0.2.1

psycopg2==2.5.2
django-cms==2.4.3
django-userprofiles==0.4
easy-thumbnails==1.4
-e git+git://github.com/Markush2010/django-bootstrap3.git@develop#egg=django-bootstrap3
-e git+git@git.sinnwerkstatt.com:cosinnus/cosinnus-core.git@ecobasa#egg=cosinnus
-e git+git@git.sinnwerkstatt.com:cosinnus/cosinnus-message.git@ecobasa#egg=cosinnus-message
-e git+git@git.sinnwerkstatt.com:cosinnus/cosinnus-todo.git@37cc10b7c490e58e42af5c1bcd659ea495c0b043#egg=cosinnus-todo
-e git+git@git.sinnwerkstatt.com:cosinnus/cosinnus-etherpad.git@ecobasa#egg=cosinnus-etherpad
-e git+git@git.sinnwerkstatt.com:cosinnus/cosinnus-event.git@3040a052a255c1014779449200647d633e9f5bbb#egg=cosinnus-event
-e git+git@git.sinnwerkstatt.com:cosinnus/cosinnus-note.git@ecobasa#egg=cosinnus-note
-e git+git://github.com/sinnwerkstatt/django-osm-field.git@master#egg=django-osm-field
django-sekizai==0.7
django-suit==0.2.8
django-filer==0.9.5
cmsplugin-filer==0.9.5
django-recaptcha==0.0.9
djangocms-text-ckeditor<2
Whoosh==2.6.0
django-haystack==2.1.0
django-honeypot==0.4.0
# workaround to missing dependencies in cosinnus:
django-tinymce==1.5.2
-e git+git://github.com/bmispelon/django-multiform.git@ccc958db8daeba6044ec18c5ff6b158bc6a3d135#egg=django-multiform
django-widget-tweaks==1.3.0
